Enabled tx broadcast to p2p node
