+++ examples/websocket_sample/env.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+export RUN="WatchAddrReceived"
+export ENDPOINT=ws://localhost:9099/ws
+export ADDR=tb1q8afqk8p2en03m0e9dtpyxhyk9emuzr3u2hey04
+export START=0
+export END=0
+export MEMPOOL=false
+export HEX=
